feat: implement interactive reconcilation
This change allows the user to define rules for automatic reconcilation
of the new transactions. If the algorithm does not find a matching rule,
it prompts the user for action, i.e. to add a new rule into the config
file and reload it, or to leave the transaction unreconciled.