feat!: revert to appending a file instead of insertion
With tools like Fava, it does not make sense to be to pedantic about
the position of a new entry in the Beancount input file. Fava provides a
view of entries sorted according to their accounts. And appending to a
file removes the complexity of finding a mark proper lineno prior
insertions.

BREAKING CHANGE: Custom directives 'beanlerk-mark' lose their meaning.
Beanclerk appends entries to a file instead of insertions.